chore(editor-web-component): Implement feedback of Kulla, simplify use-shadow-root hook and rename
