+++ packages/editor/src/core/helpers/use-shadow-root.ts
@@ -7,14 +7,11 @@ export function useShadowRoot(containerRef: React.RefObject<HTMLElement>) {
     // Should be fixed in React 19, hopefully. If not for the timeout,
     // getRootNode() will not consistently return the shadow root...
     const timeout = setTimeout(() => {
-      if (
-        containerRef.current &&
-        (containerRef.current.shadowRoot ||
-          containerRef.current.getRootNode() instanceof ShadowRoot)
-      ) {
+      const possibleShadowRoot = containerRef.current?.getRootNode()
+      if (containerRef.current && possibleShadowRoot instanceof ShadowRoot) {
         // containerRef..current.shadowRoot is always null which is why we
         // use getRootNode.
-        setShadowRoot(containerRef.current.getRootNode() as ShadowRoot)
+        setShadowRoot(possibleShadowRoot)
       }
     }, 1)
